Update RDPShell.cs
shorter loop and switch to Enter key detection.  Keypresses are cleared between the context switch from the logon screen to the user session, so holding a key won't be detected.  hopefully "double pressing" enter will be detected, but I worry that pushing it again too fast may still be in the sign-in screen context.