Update RDPShell.cs
Changed watch routine looking for dialog boxes to look for any window with the exact title "Remote Desktop Connection".  We might close on the "Connecting..." status window, but if the user locks the PC while it's connecting it's not unreasonable to assume they changed their mind and didn't want to connect anyway.  An active RDP session should start with that title but then have info about the connection as well.